get audio 5

+++ scripts/identify.js
@@ -44,7 +44,7 @@ define(function(require) {
 					volume.connect(recorder);
 
                     recorder.port.onmessage = (e) => {
-                        const samples = new Float32Array(e.data);
+                        const samples = new Int16Array(e.data);
 						chunks.push(samples); 
 					}
